refactor: consolidate fallback helpers into agno.fallback for agent+team reuse
Move fallback call helpers from agent/_fallback.py into agno/fallback.py
alongside FallbackConfig. Functions now accept (model, fallback_config)
instead of agent, making them reusable by teams.
