+++ libs/agno/agno/agent/_response.py
@@ -23,7 +23,7 @@ from pydantic import BaseModel
 if TYPE_CHECKING:
     from agno.agent.agent import Agent
 
-from agno.agent._fallback import acall_model_stream_with_fallback, call_model_stream_with_fallback
+from agno.fallback import acall_model_stream_with_fallback, call_model_stream_with_fallback
 from agno.media import Audio
 from agno.models.base import Model
 from agno.models.message import Message
@@ -1064,7 +1064,8 @@ def handle_model_response_stream(
         stream_model_response = False
 
     for model_response_event in call_model_stream_with_fallback(
-        agent,
+        agent.model,
+        agent.fallback_config,
         messages=run_messages.messages,
         response_format=response_format,
         tools=tools,
@@ -1214,7 +1215,8 @@ async def ahandle_model_response_stream(
         stream_model_response = False
 
     model_response_stream = acall_model_stream_with_fallback(
-        agent,
+        agent.model,
+        agent.fallback_config,
         messages=run_messages.messages,
         response_format=response_format,
         tools=tools,
